[Upd] currency is up to date
