--- src-tauri/src/api/mod.rs
@@ -1,3 +1,0 @@
-pub mod weather;
-pub mod currency;
-pub mod transports;
