+++ src-tauri/src/currency/mod.rs
@@ -1,0 +1,5 @@
+mod api;
+pub mod models;
+mod database;
+
+pub mod commands;
