--- src-tauri/src/models/mod.rs
@@ -1,3 +1,0 @@
-pub mod weather;
-pub mod currency;
-pub mod transports;
